refactor(utils): rehome redis_config to foundation layer (VAL-P4A-003)
aragora/utils/redis_cache.py imported aragora.server.redis_config lazily, an
upward foundation->interface layer violation. The Redis connection-pool surface
only depends on aragora.exceptions (a foundation sibling), so move it down to
aragora.utils.redis_config and leave a DeprecationWarning re-export shim at
aragora.server.redis_config for back-compat.

- redis_cache._get_redis now imports from aragora.utils.redis_config (T1 PASS)
- server/redis_config.py: pure re-export shim, warns on import (old path token)
- repoint tests that patch the impl/logger-name to the new module path
- shrink the now-zero 'aragora.utils -> aragora.server' import-contracts
  baseline edge (redis_cache was its sole source)
