+++ aragora/utils/redis_cache.py
@@ -110,7 +110,7 @@ class RedisTTLCache(Generic[T]):
             return self._redis
 
         try:
-            from aragora.server.redis_config import get_redis_client
+            from aragora.utils.redis_config import get_redis_client
 
             self._redis = get_redis_client()
             self._redis_checked = True
