+++ tests/server/test_redis_config_validation.py
@@ -24,7 +24,7 @@ def _reset_and_get_pool(env_overrides: dict[str, str] | None = None):
     The Redis module uses module-level state that must be reset between tests.
     Returns the pool and availability status after initialization.
     """
-    import aragora.server.redis_config as redis_mod
+    import aragora.utils.redis_config as redis_mod
 
     # Reset module state
     redis_mod.reset_redis_state()
@@ -93,19 +93,19 @@ class TestMaxConnectionsValidation:
 
     def test_max_connections_zero_clamped_to_1(self, caplog):
         """max_connections of 0 is clamped to 1."""
-        with caplog.at_level(logging.WARNING, logger="aragora.server.redis_config"):
+        with caplog.at_level(logging.WARNING, logger="aragora.utils.redis_config"):
             max_connections = _reset_and_get_pool({"ARAGORA_REDIS_MAX_CONNECTIONS": "0"})
         assert max_connections == 1
 
     def test_max_connections_negative_clamped_to_1(self, caplog):
         """max_connections of -5 is clamped to 1."""
-        with caplog.at_level(logging.WARNING, logger="aragora.server.redis_config"):
+        with caplog.at_level(logging.WARNING, logger="aragora.utils.redis_config"):
             max_connections = _reset_and_get_pool({"ARAGORA_REDIS_MAX_CONNECTIONS": "-5"})
         assert max_connections == 1
 
     def test_max_connections_exceeding_10000_capped(self, caplog):
         """max_connections > 10000 is capped to 10000."""
-        with caplog.at_level(logging.WARNING, logger="aragora.server.redis_config"):
+        with caplog.at_level(logging.WARNING, logger="aragora.utils.redis_config"):
             max_connections = _reset_and_get_pool({"ARAGORA_REDIS_MAX_CONNECTIONS": "15000"})
         assert max_connections == 10000
 
@@ -115,12 +115,12 @@ class TestMaxConnectionsWarningLogs:
 
     def test_clamping_to_1_logs_warning(self, caplog):
         """Clamping max_connections to 1 logs a warning."""
-        with caplog.at_level(logging.WARNING, logger="aragora.server.redis_config"):
+        with caplog.at_level(logging.WARNING, logger="aragora.utils.redis_config"):
             _reset_and_get_pool({"ARAGORA_REDIS_MAX_CONNECTIONS": "0"})
         assert any("clamping to 1" in record.message for record in caplog.records)
 
     def test_capping_at_10000_logs_warning(self, caplog):
         """Capping max_connections to 10000 logs a warning."""
-        with caplog.at_level(logging.WARNING, logger="aragora.server.redis_config"):
+        with caplog.at_level(logging.WARNING, logger="aragora.utils.redis_config"):
             _reset_and_get_pool({"ARAGORA_REDIS_MAX_CONNECTIONS": "20000"})
         assert any("capping at 10000" in record.message for record in caplog.records)
